branch:	HACKT-00-01-04-main-00-81-39-template-02 rev:	HACKT-00-01-04-main-00-81-39-template-02-03
  No source revision, only test updates:
	the few 'regressions' have been updated from create-fail
	to create-pass and alloc-fail cases.  No regressions remain.
  Next: continue implementation of template_type_completion
	and add test cases.

+++ test/update-allocstderr.sh
@@ -23,7 +23,17 @@ done
 
 for i
 do
-	cp -f $i.allocdump-filter $srcdir/$i.allocstderr
-	echo "Updated $srcdir/$i.allocstderr"
+	if test -f $i.allocdump-filter
+	then
+		cp -f $i.allocdump-filter $srcdir/$i.allocstderr
+		echo "Updated $srcdir/$i.allocstderr"
+	elif test -f $i.allocfaildump
+	then
+		cp -f $i.allocfaildump $srcdir/$i.allocstderr
+		echo "Updated $srcdir/$i.allocstderr"
+	else
+		echo "$i.allocdump-filter or $i.allocfaildump required but missing."
+	fi
 done
+
